refactor(size1Dcomponent): Change the type of the variable
Change it to size_t because it can't be negative.

+++ Client/src/Component/Size1DComponent.cpp
@@ -7,5 +7,5 @@
 
 #include "Size1DComponent.hpp"
 
-Size1DComponent::Size1DComponent(int size) :
+Size1DComponent::Size1DComponent(size_t size) :
     AComponent("Size1DComponent"), size(size) {}
